+++ pkg/main.go
@@ -13,7 +13,7 @@ import (
 func main() {
 	backend.Logger.Info("Starting VictoriaLogs datasource backend ...")
 
-	if err := datasource.Manage("victorialogs-datasource-http-backend", plugin.NewDatasource, datasource.ManageOpts{}); err != nil {
+	if err := datasource.Manage("victoriametrics-logs-datasource", plugin.NewDatasource, datasource.ManageOpts{}); err != nil {
 		log.DefaultLogger.Error("Failed to process VictoriaLogs datasource backend: %s", err.Error())
 		os.Exit(1)
 	}